HTTP Client Test Consolidation, Remove Boost.Test
This is the simplification of the HTTP client test suite which serves as
a guide for getting other tests migrated to use gtest. Moving forward
the idea is to accomplish the same for all of the tests in cpp-netlib:

* Consolidate logically grouped tests into a single binary.
* Remove all traces of dependency to the Boost.Test framework.
* Get all tests to a "green" state again.

This paves the way for implementing mocks for some internal classes that
we want to be able to change. In the meantime simplifying the test suite
buys us more in the long run as we add more and more tests to the
library.